Rename promises utils.
